fc_rank/requirements.txt
certifi==2018.4.16
chardet==3.0.4
docopt==0.6.2
idna==2.7
jedi==0.12.1
lxml==4.2.3
parso==0.3.0
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptpython==0.41
pyaml==17.12.1
Pygments==2.2.0
PyYAML==3.12
requests==2.19.1
six==1.11.0
urllib3==1.23
wcwidth==0.1.7
